mm: Fix incorrect type conversion for size during dma allocation
CVE-2017-9725

This was found during userspace fuzzing test when a large size
allocation is made from ion

+++ drivers/base/dma-contiguous.c
@@ -581,7 +581,7 @@ static void clear_cma_bitmap(struct cma *cma, unsigned long pfn, int count)
  * global one. Requires architecture specific get_dev_cma_area() helper
  * function.
  */
-unsigned long dma_alloc_from_contiguous(struct device *dev, int count,
+unsigned long dma_alloc_from_contiguous(struct device *dev, size_t count,
 				       unsigned int align)
 {
 	unsigned long mask, pfn = 0, pageno, start = 0;
